Changed some of the simulator api to use the repository. Many methods still need to be changed
